+++ sdk/Sdk/DataTypeEnum.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+namespace Microsoft.Azure.Functions.Worker.Sdk
+{
+    internal enum DataType
+    {
+        Undefined,
+        Binary,
+        String,
+        Stream
+    }
+}